# lib/api-zod/src/generated/types/broadcasterPair.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * USALB Radio live broadcasting and station management API
 * OpenAPI spec version: 0.1.0
 */

export interface BroadcasterPair {
  stationName: string;
  deviceId: string;
  displayName: string;
  publishToken: string;
  publishEndpoint: string;
  publicStreamEndpoint: string;
  heartbeatEndpoint: string;
  telemetryEndpoint: string;
  intentEndpoint: string;
  commandsEndpoint: string;
  format: string;
}