# lib/api-zod/src/generated/types/broadcasterTelemetryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * USALB Radio live broadcasting and station management API
 * OpenAPI spec version: 0.1.0
 */

export interface BroadcasterTelemetryInput {
  /** @maxLength 64 */
  status?: string;
  /**
     * @minimum 0
     * @nullable
     */
  bitrateKbps?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  sampleRate?: number | null;
  /**
     * @maxLength 120
     * @nullable
     */
  contentType?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  bytesSent?: number | null;
  /**
     * @maxLength 500
     * @nullable
     */
  detail?: string | null;
}